+++ streetSoccerManager-Backend/streetSoccerManager/src/main/java/soccer/game/streetSoccerManager/repository/repositoryInterfaces/IFormationRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface IFormationRepository {
     List<Formation> getAll();
-    Formation get(int id);
-    Boolean delete(int id);
+    Formation get(Long id);
+    Boolean delete(Long id);
     Boolean add(Formation formation);
     Boolean update(Formation formation);
 }
